hdfc/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.4
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
exceptiongroup==1.1.0
httpie==3.2.1
idna==3.4
iniconfig==1.1.1
jwcrypto==1.4.2
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
npm==0.1.1
numpy==1.24.1
oauthlib==3.2.2
optional-django==0.1.0
packaging==22.0
pandas==1.5.2
pluggy==1.0.0
psycopg2==2.9.5
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pymongo==3.12.3
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.1
six==1.16.0
sqlparse==0.2.4
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
wrapt==1.15.0
